# docs/api/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/upcoming/{id}":{"get":{"tags":["upcoming-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Upcoming"}}}}}},"put":{"tags":["upcoming-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Upcoming"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Upcoming"}}}}}},"delete":{"tags":["upcoming-controller"],"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/tournaments/{id}":{"get":{"tags":["tournament-controller"],"operationId":"getById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Tournament"}}}}}},"put":{"tags":["tournament-controller"],"operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tournament"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Tournament"}}}}}},"delete":{"tags":["tournament-controller"],"operationId":"delete_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/mediazone_news/{id}":{"get":{"tags":["mediazone-news-controller"],"operationId":"getById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MediazoneNews"}}}}}},"put":{"tags":["mediazone-news-controller"],"operationId":"update_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MediazoneNews"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MediazoneNews"}}}}}},"delete":{"tags":["mediazone-news-controller"],"operationId":"delete_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/mediazone_gallery/{id}":{"get":{"tags":["mediazone-gallery-controller"],"operationId":"getById_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MediazoneGallery"}}}}}},"put":{"tags":["mediazone-gallery-controller"],"operationId":"update_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MediazoneGallery"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MediazoneGallery"}}}}}},"delete":{"tags":["mediazone-gallery-controller"],"operationId":"delete_3","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/mediazone/{id}":{"get":{"tags":["mediazone-controller"],"operationId":"getById_4","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Mediazone"}}}}}},"put":{"tags":["mediazone-controller"],"operationId":"update_4","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Mediazone"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Mediazone"}}}}}},"delete":{"tags":["mediazone-controller"],"operationId":"delete_4","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/account/profile":{"get":{"tags":["account-controller"],"operationId":"getProfile","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Account"}}}}}},"put":{"tags":["account-controller"],"operationId":"updateProfile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAccountRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Account"}}}}}}},"/api/account/password":{"put":{"tags":["account-controller"],"operationId":"changePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/upcoming":{"get":{"tags":["upcoming-controller"],"operationId":"getAll","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Upcoming"}}}}}}},"post":{"tags":["upcoming-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Upcoming"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Upcoming"}}}}}}},"/api/tournaments":{"get":{"tags":["tournament-controller"],"operationId":"getAll_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Tournament"}}}}}}},"post":{"tags":["tournament-controller"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tournament"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Tournament"}}}}}}},"/api/reviews":{"get":{"tags":["review-controller"],"operationId":"getAll_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Review"}}}}}}},"post":{"tags":["review-controller"],"operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Review"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Review"}}}}}}},"/api/mediazone_news":{"get":{"tags":["mediazone-news-controller"],"operationId":"getAll_3","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MediazoneNews"}}}}}}},"post":{"tags":["mediazone-news-controller"],"operationId":"create_3","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MediazoneNews"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MediazoneNews"}}}}}}},"/api/mediazone_gallery":{"get":{"tags":["mediazone-gallery-controller"],"operationId":"getAll_4","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MediazoneGallery"}}}}}}},"post":{"tags":["mediazone-gallery-controller"],"operationId":"create_4","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MediazoneGallery"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MediazoneGallery"}}}}}}},"/api/mediazone":{"get":{"tags":["mediazone-controller"],"operationId":"getAll_5","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Mediazone"}}}}}}},"post":{"tags":["mediazone-controller"],"operationId":"create_5","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Mediazone"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Mediazone"}}}}}}},"/api/account/register":{"post":{"tags":["account-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAccountRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Account"}}}}}}},"/api/account/login":{"post":{"tags":["account-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Account"}}}}}}},"/api/reviews/{id}":{"get":{"tags":["review-controller"],"operationId":"getById_5","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Review"}}}}}},"delete":{"tags":["review-controller"],"operationId":"delete_5","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/account/{id}":{"delete":{"tags":["account-controller"],"operationId":"deleteAccount","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"MapText":{"type":"object","properties":{"teams":{"type":"array","items":{"type":"string"}},"format":{"type":"array","items":{"type":"string"}},"maps":{"type":"array","items":{"type":"string","enum":["MIRAGE","INFERNO","OVERPASS","NUKE","VERTIGO"]}}}},"Upcoming":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"createdAt":{"type":"string","format":"date-time"},"city":{"type":"string"},"title":{"type":"string"},"price":{"type":"string"},"tournamentType":{"type":"string","enum":["ONLINE","OFFLINE"]},"cardText":{"$ref":"#/components/schemas/MapText"}},"required":["title"]},"Tournament":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"thumb":{"type":"string"},"title":{"type":"string","minLength":1},"formatTournament":{"type":"string"},"prize":{"type":"string"},"data":{"type":"string","format":"date"},"status":{"type":"string","enum":["REGISTRATION","STARTED","ENDED"]},"slots":{"type":"string"},"type":{"type":"string","enum":["ONLINE","OFFLINE"]},"mix":{"type":"boolean"}},"required":["data","status","title"]},"MediazoneNews":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"date":{"type":"string","format":"date"},"title":{"type":"string","minLength":1},"text":{"type":"string"},"filtres":{"type":"array","items":{"type":"string"}}},"required":["title"]},"MediazoneGallery":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"title":{"type":"string","minLength":1},"imageUrl":{"type":"string","minLength":1},"description":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}},"required":["imageUrl","title"]},"Mediazone":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"imgs":{"type":"array","items":{"type":"string"}}}},"UpdateAccountRequest":{"type":"object","properties":{"username":{"type":"string"},"email":{"type":"string"},"profileImage":{"type":"string"}}},"Account":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"authAccount":{"$ref":"#/components/schemas/AuthAccount"}}},"AuthAccount":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"username":{"type":"string"},"email":{"type":"string"},"passwordHash":{"type":"string"}}},"ChangePasswordRequest":{"type":"object","properties":{"oldPassword":{"type":"string","minLength":1},"newPassword":{"type":"string","minLength":1}},"required":["newPassword","oldPassword"]},"Review":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"author":{"type":"string","minLength":1},"text":{"type":"string","minLength":1},"avatarImage":{"type":"string"},"role":{"type":"string","enum":["PARTICIPANT","ORGANIZER","ADMIN"]},"createdAt":{"type":"string","format":"date-time"}},"required":["author","text"]},"CreateAccountRequest":{"type":"object","properties":{"username":{"type":"string","minLength":1},"email":{"type":"string","format":"email","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","password","username"]},"LoginRequest":{"type":"object","properties":{"login":{"type":"string","minLength":1},"password":{"type":"string","minLength":1}},"required":["login","password"]}}}}